+++ PeriodicTableDictionary.cpp
@@ -80,7 +80,7 @@ int PeriodicTable :: elementPos() const {
     return (afterCursor->atomicNumber);
 }
 
-long PeriodicTable :: elmentWeight () const {
+float PeriodicTable :: elmentWeight () const {
     return (afterCursor->atomicMass);
 }
